Move the 4x4 functional tests to a separate test case
